+++ src/scene/shader-lib/glsl/chunks/gsplat/vert/gsplatModify.js
