+++ src/scene/shader-lib/wgsl/chunks/gsplat/vert/gsplatCommon.js
@@ -1,7 +1,7 @@
 export default /* wgsl */`
 #include "gsplatHelpersVS"
 #include "gsplatCustomizeVS"
-#include "gsplatModifySplatVS"
+#include "gsplatModifyVS"
 
 #include "gsplatStructsVS"
 #include "gsplatEvalSHVS"
