+++ src/scene/shader-lib/wgsl/chunks/gsplat/vert/gsplatModify.js
